requirements.txt
asgiref==3.5.2
Django==4.1.2
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.3
